Fix a typo (VSCode -> VS Code) in README.md
Repository description and "Extensions and the Marketplace" `README.md`
paragraph refers to the upstream editor as "VS Code", while other parts of the
README refer to it as "VSCode".

Upstream refers to it as "VS Code", this commit standardizes on that specific
spelling (e.g. see https://github.com/microsoft/vscode#related-projects).